move package manager from event listener
