+++ api/src/main/java/com/pfplaybackend/api/config/websocket/event/listener/UnsubscriptionEventListener.java
@@ -1,7 +1,7 @@
 package com.pfplaybackend.api.config.websocket.event.listener;
 
 import com.pfplaybackend.api.common.exception.ExceptionCreator;
-import com.pfplaybackend.api.config.websocket.event.manager.SessionCacheManager;
+import com.pfplaybackend.api.config.websocket.manager.SessionCacheManager;
 import com.pfplaybackend.api.config.websocket.exception.SessionException;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
